fixed deleting of authors

+++ library/author/views.py
@@ -2,6 +2,7 @@
 from django.shortcuts import render, redirect
 from django.contrib import messages
 from author.models import Author
+from book.models import Book
 
 def show_all_authors(request):
     if not request.user.is_authenticated:
@@ -50,18 +51,24 @@ def delete_author(request, author_id):
         return redirect('auth:dashboard')
     
     if request.method == 'POST':
+        books = Book.objects.all()
+        authors_ids = set([author.id for book in books for author in book.authors.all()])
         author = Author.get_by_id(author_id=author_id)
-        
+
         if author is None:
             messages.error(request, "Perhaps archives are incomplete. We can't delete what we don't know")
             return redirect('author:authors')
+
+        if author.id not in authors_ids:
+            deleted = Author.delete_by_id(author_id=author_id)
         
-        deleted = Author.delete_by_id(author_id=author_id)
+            if not deleted:
+                messages.error(request, "Something went wrong. The author wasn't deleted")
+                return redirect('author:authors')
         
-        if not deleted:
-            messages.error(request, "Something went wrong. The author wasn't deleted")
             return redirect('author:authors')
-        
-        return redirect('author:authors')
+        else:
+            messages.error(request, "Can't delete this author")
+            return redirect('author:authors')
     
     return redirect('author:authors')
